requirements.txt
asgiref==3.2.7
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
cryptography==3.1.1
defusedxml==0.6.0
Django==3.0
django-allauth==0.42.0
djangorestframework==3.11.1
idna==2.10
oauthlib==3.1.0
Pillow==7.0.0
pycparser==2.20
PyJWT==1.7.1
python3-openid==3.2.0
pytz==2019.3
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.3.1
urllib3==1.25.10
